@@ -2,60 +2,59 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace BuildingBlocks.Domain
+namespace BuildingBlocks.Domain;
+
+public abstract class ValueObject
 {
-    public abstract class ValueObject
+    protected static bool EqualOperator(ValueObject? left, ValueObject? right)
     {
-        protected static bool EqualOperator(ValueObject? left, ValueObject? right)
+        if (ReferenceEquals(left, null) ^ ReferenceEquals(right, null))
         {
-            if (ReferenceEquals(left, null) ^ ReferenceEquals(right, null))
-            {
-                return false;
-            }
-            return ReferenceEquals(left, null) || left.Equals(right);
+            return false;
+        }
+        return ReferenceEquals(left, null) || left.Equals(right);
+    }
+
+    protected static bool NotEqualOperator(ValueObject? left, ValueObject? right)
+    {
+        return !(EqualOperator(left, right));
+    }
+
+
+    public static bool operator ==(ValueObject? left, ValueObject? right)
+    {
+        return EqualOperator(left, right);
+    }
+
+    public static bool operator !=(ValueObject? left, ValueObject? right)
+    {
+        return NotEqualOperator(left, right);
+    }
+
+
+    protected abstract IEnumerable<object?> GetEqualityComponents();
+
+    public override bool Equals(object? obj)
+    {
+        if (obj is null || obj.GetType() != GetType())
+        {
+            return false;
         }
 
-        protected static bool NotEqualOperator(ValueObject? left, ValueObject? right)
-        {
-            return !(EqualOperator(left, right));
-        }
+        var other = (ValueObject)obj;
 
+        return this.GetEqualityComponents().SequenceEqual(other.GetEqualityComponents());
+    }
 
-        public static bool operator ==(ValueObject? left, ValueObject? right)
-        {
-            return EqualOperator(left, right);
-        }
+    public override int GetHashCode()
+    {
+        return GetEqualityComponents()
+            .Select(x => x != null ? x.GetHashCode() : 0)
+            .Aggregate((x, y) => x ^ y);
+    }
 
-        public static bool operator !=(ValueObject? left, ValueObject? right)
-        {
-            return NotEqualOperator(left, right);
-        }
-
-
-        protected abstract IEnumerable<object?> GetEqualityComponents();
-
-        public override bool Equals(object? obj)
-        {
-            if (obj is null || obj.GetType() != GetType())
-            {
-                return false;
-            }
-
-            var other = (ValueObject)obj;
-
-            return this.GetEqualityComponents().SequenceEqual(other.GetEqualityComponents());
-        }
-
-        public override int GetHashCode()
-        {
-            return GetEqualityComponents()
-                .Select(x => x != null ? x.GetHashCode() : 0)
-                .Aggregate((x, y) => x ^ y);
-        }
-
-        public ValueObject? GetCopy()
-        {
-            return MemberwiseClone() as ValueObject;
-        }
+    public ValueObject? GetCopy()
+    {
+        return MemberwiseClone() as ValueObject;
     }
 }
