@@ -1,41 +1,37 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using StorekeeperAssistant.DataAccess.EntityConfigurations;
-using StorekeeperAssistant.Domain.InventoryItemAggregate;
-using StorekeeperAssistant.Domain.MovingAggregate;
-using StorekeeperAssistant.Domain.MovingAggregate.WarehouseInventoryItems;
-using StorekeeperAssistant.Domain.WarehouseAggregate;
+using StorekeeperAssistant.Domain.InventoryItems;
+using StorekeeperAssistant.Domain.Movings;
+using StorekeeperAssistant.Domain.Movings.WarehouseInventoryItems;
+using StorekeeperAssistant.Domain.Warehouses;
 using System;
 using System.Collections.Generic;
+using System.Reflection;
 
-namespace StorekeeperAssistant.DataAccess
+namespace StorekeeperAssistant.DataAccess;
+
+public sealed class AppDbContext : DbContext
 {
-    public class AppDbContext : DbContext
-    {
 
 #nullable disable
-        public DbSet<Moving> Movings { get; set; }
-        public DbSet<Warehouse> Warehouses { get; set; }
-        public DbSet<InventoryItem> InventoryItems { get; set; }
-        public DbSet<WarehouseInventoryItem> WarehouseInventoryItems { get; set; }
+    public DbSet<Moving> Movings { get; set; }
+    public DbSet<Warehouse> Warehouses { get; set; }
+    public DbSet<InventoryItem> InventoryItems { get; set; }
+    public DbSet<WarehouseInventoryItem> WarehouseInventoryItems { get; set; }
 #nullable restore
 
-        public AppDbContext(DbContextOptions<AppDbContext> options)
-            : base(options)
-        {
-        }
+    public AppDbContext(DbContextOptions<AppDbContext> options)
+        : base(options)
+    {
+    }
 
-        internal IEnumerable<object> OrderByDescending(Func<object, object> p)
-        {
-            throw new NotImplementedException();
-        }
+    internal IEnumerable<object> OrderByDescending(Func<object, object> p)
+    {
+        throw new NotImplementedException();
+    }
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.ApplyConfiguration(new MovingEntityTypeConfiguration());
-            modelBuilder.ApplyConfiguration(new WarehouseEntityTypeConfiguration());
-            modelBuilder.ApplyConfiguration(new InventoryItemEntityTypeConfiguration());
-            modelBuilder.ApplyConfiguration(new MovingDetailEntityTypeConfiguration());
-            modelBuilder.ApplyConfiguration(new WarehouseInventoryItemEntityTypeConfiguration());
-        }
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        //warning!!! with this approach you have no control over the order in which configurations are applied!!!
+        modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());
     }
 }
