@@ -1,11 +1,10 @@
-﻿using StorekeeperAssistant.Domain.InventoryItemAggregate;
+﻿using StorekeeperAssistant.Domain.InventoryItems;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace StorekeeperAssistant.UseCases.Interfaces
+namespace StorekeeperAssistant.UseCases.Interfaces;
+
+public interface IInventoryItemRepository
 {
-    public interface IInventoryItemRepository
-    {
-        Task<IEnumerable<InventoryItem>> GetByIds(IEnumerable<InventoryItemId> inventoryItemIds);
-    }
+    Task<IEnumerable<InventoryItem>> GetByIds(IEnumerable<InventoryItemId> inventoryItemIds);
 }
