@@ -1,12 +1,11 @@
-﻿using StorekeeperAssistant.Domain.InventoryItemAggregate;
-using StorekeeperAssistant.Domain.MovingAggregate.WarehouseInventoryItems;
-using StorekeeperAssistant.Domain.WarehouseAggregate;
+﻿using StorekeeperAssistant.Domain.InventoryItems;
+using StorekeeperAssistant.Domain.Movings.WarehouseInventoryItems;
+using StorekeeperAssistant.Domain.Warehouses;
 using System.Threading.Tasks;
 
-namespace StorekeeperAssistant.UseCases.Interfaces
+namespace StorekeeperAssistant.UseCases.Interfaces;
+
+public interface IWarehouseInventoryItemRepository
 {
-    public interface IWarehouseInventoryItemRepository
-    {
-        Task<WarehouseInventoryItem> Get(WarehouseId warehouseId, InventoryItemId inventoryItemId);
-    }
+    Task<WarehouseInventoryItem?> Get(WarehouseId warehouseId, InventoryItemId inventoryItemId);
 }
