@@ -1,13 +1,11 @@
-﻿using StorekeeperAssistant.Domain.WarehouseAggregate;
-using System;
+﻿using StorekeeperAssistant.Domain.Warehouses;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace StorekeeperAssistant.UseCases.Interfaces
+namespace StorekeeperAssistant.UseCases.Interfaces;
+
+public interface IWarehouseRepository
 {
-    public interface IWarehouseRepository
-    {
-        Task<IEnumerable<Warehouse>> GetByIds(IEnumerable<WarehouseId> warehouseIds);
-        Task<Warehouse?> GetById(WarehouseId warehouseId);
-    }
+    Task<IEnumerable<Warehouse>> GetByIds(IEnumerable<WarehouseId> warehouseIds);
+    Task<Warehouse?> GetById(WarehouseId warehouseId);
 }
