@@ -1,7 +1,37 @@
-﻿using MediatR;
+﻿using BuildingBlocks.UseCases;
+using Dapper;
+using MediatR;
 using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
 
-namespace StorekeeperAssistant.UseCases.InventoryItems.Queries.GetInventoryItems
+namespace StorekeeperAssistant.UseCases.InventoryItems.Queries.GetInventoryItems;
+
+public sealed record GetInventoryItemsQuery(): IRequest<IEnumerable<InventoryItemDto>>;
+
+public sealed class GetInventoryItemsQueryHandler : IRequestHandler<GetInventoryItemsQuery, IEnumerable<InventoryItemDto>>
 {
-    public record GetInventoryItemsQuery(): IRequest<IEnumerable<InventoryItemDto>>;
+    private readonly ISqlConnectionFactory _sqlConnectionFactory;
+
+    public GetInventoryItemsQueryHandler(ISqlConnectionFactory sqlConnectionFactory)
+    {
+        _sqlConnectionFactory = sqlConnectionFactory;
+    }
+
+    public async Task<IEnumerable<InventoryItemDto>> Handle(GetInventoryItemsQuery request, CancellationToken cancellationToken)
+    {
+        var db = _sqlConnectionFactory.GetOpenConnection();
+
+        return await db.QueryAsync<InventoryItemDto>(
+            "SELECT " +
+            "[Id], " +
+            "[Name] " +
+            "FROM [InventoryItems] " +
+            "WHERE [IsDeleted] = @IsDeleted " +
+            "ORDER BY [Name]",
+            new
+            {
+                IsDeleted = false
+            });
+    }
 }
