@@ -1,10 +1,9 @@
 ﻿using System;
 
-namespace StorekeeperAssistant.UseCases
+namespace StorekeeperAssistant.UseCases;
+
+public sealed class InventoryItemDto
 {
-    public class InventoryItemDto
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; } = default!;
-    }
+    public Guid Id { get; set; }
+    public string Name { get; set; } = default!;
 }
