@@ -1,100 +1,101 @@
 ﻿using System;
 using Microsoft.EntityFrameworkCore.Migrations;
 
-namespace StorekeeperAssistant.Web.Migrations
+namespace StorekeeperAssistant.Web.Migrations;
+
+#nullable disable
+
+public partial class init : Migration
 {
-    public partial class init : Migration
+    protected override void Up(MigrationBuilder migrationBuilder)
     {
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.CreateTable(
-                name: "InventoryItems",
-                columns: table => new
-                {
-                    Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    IsDeleted = table.Column<bool>(type: "bit", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_InventoryItems", x => x.Id);
-                });
+        migrationBuilder.CreateTable(
+            name: "InventoryItems",
+            columns: table => new
+            {
+                Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
+                IsDeleted = table.Column<bool>(type: "bit", nullable: false)
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_InventoryItems", x => x.Id);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "MovingDetails",
-                columns: table => new
-                {
-                    Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    Count = table.Column<int>(type: "int", nullable: false),
-                    MovingId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    InventoryItemId = table.Column<Guid>(type: "uniqueidentifier", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_MovingDetails", x => x.Id);
-                });
+        migrationBuilder.CreateTable(
+            name: "MovingDetails",
+            columns: table => new
+            {
+                Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                Count = table.Column<int>(type: "int", nullable: false),
+                MovingId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                InventoryItemId = table.Column<Guid>(type: "uniqueidentifier", nullable: false)
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_MovingDetails", x => x.Id);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "Movings",
-                columns: table => new
-                {
-                    Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    TransferDate = table.Column<DateTime>(type: "datetime2", nullable: false),
-                    DepartureWarehouseId = table.Column<Guid>(type: "uniqueidentifier", nullable: true),
-                    ArrivalWarehouseId = table.Column<Guid>(type: "uniqueidentifier", nullable: true),
-                    IsDeleted = table.Column<bool>(type: "bit", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_Movings", x => x.Id);
-                });
+        migrationBuilder.CreateTable(
+            name: "Movings",
+            columns: table => new
+            {
+                Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                TransferDate = table.Column<DateTime>(type: "datetime2", nullable: false),
+                DepartureWarehouseId = table.Column<Guid>(type: "uniqueidentifier", nullable: true),
+                ArrivalWarehouseId = table.Column<Guid>(type: "uniqueidentifier", nullable: true),
+                IsDeleted = table.Column<bool>(type: "bit", nullable: false)
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_Movings", x => x.Id);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "WarehouseInventoryItems",
-                columns: table => new
-                {
-                    Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    Count = table.Column<int>(type: "int", nullable: false),
-                    Date = table.Column<DateTime>(type: "datetime2", nullable: false),
-                    WarehouseId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    InventoryItemId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    MovingId = table.Column<Guid>(type: "uniqueidentifier", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_WarehouseInventoryItems", x => x.Id);
-                });
+        migrationBuilder.CreateTable(
+            name: "WarehouseInventoryItems",
+            columns: table => new
+            {
+                Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                Count = table.Column<int>(type: "int", nullable: false),
+                Date = table.Column<DateTime>(type: "datetime2", nullable: false),
+                WarehouseId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                InventoryItemId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                MovingId = table.Column<Guid>(type: "uniqueidentifier", nullable: false)
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_WarehouseInventoryItems", x => x.Id);
+            });
 
-            migrationBuilder.CreateTable(
-                name: "Warehouses",
-                columns: table => new
-                {
-                    Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    IsDeleted = table.Column<bool>(type: "bit", nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_Warehouses", x => x.Id);
-                });
-        }
+        migrationBuilder.CreateTable(
+            name: "Warehouses",
+            columns: table => new
+            {
+                Id = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
+                Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
+                IsDeleted = table.Column<bool>(type: "bit", nullable: false)
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_Warehouses", x => x.Id);
+            });
+    }
 
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropTable(
-                name: "InventoryItems");
+    protected override void Down(MigrationBuilder migrationBuilder)
+    {
+        migrationBuilder.DropTable(
+            name: "InventoryItems");
 
-            migrationBuilder.DropTable(
-                name: "MovingDetails");
+        migrationBuilder.DropTable(
+            name: "MovingDetails");
 
-            migrationBuilder.DropTable(
-                name: "Movings");
+        migrationBuilder.DropTable(
+            name: "Movings");
 
-            migrationBuilder.DropTable(
-                name: "WarehouseInventoryItems");
+        migrationBuilder.DropTable(
+            name: "WarehouseInventoryItems");
 
-            migrationBuilder.DropTable(
-                name: "Warehouses");
-        }
+        migrationBuilder.DropTable(
+            name: "Warehouses");
     }
 }
