@@ -1,9 +1,9 @@
 ﻿using BuildingBlocks.Domain;
 using System;
 
-namespace StorekeeperAssistant.Domain.InventoryItemAggregate
+namespace StorekeeperAssistant.Domain.InventoryItems
 {
-    public class InventoryItemId : EntityId
+    public sealed class InventoryItemId : EntityId
     {
         public InventoryItemId(Guid value) : base(value)
         {
